Remove postgres and mysql time specific instances
Use the JDBC 4.2 instances available by default.
