+++ modules/mysql/src/main/scala/doobie/mysql/package.scala
@@ -3,5 +3,4 @@ package doobie
 package object mysql {
 
   object implicits
-    extends JavaTimeInstances
 }
